Refactor to no longer have Test jar
