+++ integration-tests/src/test/java/org/acme/WithDefaultPlaywrightTest.java
