+++ integration-tests/src/test/java/org/acme/WithFirefoxPlaywrightTest.java
