+++ pom.xml
@@ -15,7 +15,6 @@
     <modules>
         <module>deployment</module>
         <module>runtime</module>
-        <module>testing</module>
     </modules>
     <scm>
         <connection>scm:git:git@github.com:quarkiverse/quarkus-playwright.git</connection>
