+++ runtime/src/main/java/io/quarkiverse/playwright/QuarkusPlaywrightManager.java
