+++ runtime/src/main/java/io/quarkiverse/playwright/WithPlaywright.java
